requirements/dev.txt
aniso8601==9.0.1
asgiref==3.7.2
Django==4.2.11
django-filter==24.1
django-graphql-jwt==0.4.0
factory-boy==3.3.0
Faker==24.1.0
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
promise==2.3
PyJWT==2.8.0
pytest==8.1.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
